Restore full opacity for leaf blocks
