+++ src/blocks/blocks.js
@@ -52,7 +52,7 @@ export const blocks = [
     breakable: true,
     drops: [{ id: 'leaves', count: 1 }],
     transparent: true,
-    opacity: 0.7,
+    opacity: 1.0,
     renderLayer: 'alpha',
     textures: { all: 'leaves' }
   },
